requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.0
matplotlib==3.9.3
seaborn==0.13.2
streamlit==1.41.1
sqlite3==3.45.3
python-dotenv==0.21.0
pydantic==2.8.2
pinecone-client==5.0.1
langchain==0.3.7
openai==1.52.1
semantic_router==0.0.72
fpdf==1.7.2